requirements.txt
# You need scipy, but since this project is deployed to Heroku, we can't
# add the requirement here because we need to use a custom build-pack for scipy.
# scipy==0.11.0

GDAL==1.9.1
Flask==0.8
Flask-Cache==0.8.0
gunicorn==0.17.4
locustio==0.6.2
newrelic
